requirements.txt
# NO DASHBOARD VERSION - DISCORD ALERTS ONLY
# Core System Requirements
requests>=2.31.0
numpy>=1.26.0,<2.0.0
pandas>=2.0.0
psutil>=5.9.0
click>=8.1.0
colorlog>=6.7.0
python-dotenv>=1.0.0
PyYAML>=6.0
schedule>=1.2.0
watchdog>=3.0.0
setuptools>=68.0.0
wheel>=0.40.0
pip>=23.0.0

# AI/ML Requirements - FIXED VERSIONS
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
torchaudio>=2.0.0
transformers>=4.30.0,<5.0.0
diffusers>=0.21.0,<1.0.0
accelerate>=0.20.0,<1.0.0
scikit-learn>=1.3.0,<2.0.0
opencv-python>=4.8.0,<5.0.0
Pillow>=10.1.0,<11.0.0  # NOW WORKS - no streamlit conflict
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
scipy>=1.11.4,<2.0.0
nltk>=3.8.0,<4.0.0
spacy>=3.6.0,<4.0.0
sentence-transformers>=2.2.0,<3.0.0
openai>=0.27.0,<2.0.0
huggingface-hub>=0.16.0,<1.0.0
ollama==0.1.9

# Google APIs (YouTube Integration)
google-api-python-client==2.97.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0

# Video Processing - FIXED VERSIONS
moviepy==1.0.3
imageio>=2.33.0,<3.0.0
mediapipe>=0.10.13,<0.11.0

# Web Scraping
beautifulsoup4>=4.12.0,<5.0.0
selenium>=4.10.0,<5.0.0
requests-html>=0.10.0,<1.0.0
scrapy>=2.9.0,<3.0.0
praw>=7.7.0,<8.0.0
yfinance>=0.2.0,<1.0.0
feedparser>=6.0.0,<7.0.0
webdriver-manager==3.8.6

# Audio & Voice
edge-tts==6.1.7
pydub==0.25.1
SpeechRecognition==3.10.0

# Trend Intelligence & Analytics
pytrends==4.9.2
yt-dlp==2023.7.6
textblob==0.17.1

# System Utilities
apscheduler==3.10.4
aiofiles==23.2.1

# Deep Learning (Stable Diffusion)
xformers
safetensors

# Knowledge APIs
wikipedia-api==0.6.0
wikipedia==1.4.0
aiohttp==3.8.5

# Financial APIs
pandas-datareader==0.10.0
fredapi==0.4.3
instaloader>=4.9.0,<5.0.0
ccxt==4.0.77

# RSS/News
rss-parser==0.2.4
newsapi-python>=0.2.6
newscatcherapi>=0.1.0

# Visualization - NO DASHBOARD
# streamlit==1.25.0  # Removed - Discord metrics only
plotly==5.15.0  # Still useful for generating charts for Discord

# Discord Alerts & Monitoring
gtts>=2.3.0
discord-webhook>=1.3.0  # For Discord alerts